aliens sounds added

+++ headers/Global.h
@@ -8,7 +8,7 @@ constexpr unsigned char SHIP_BULLET_SPEED = 4;
 constexpr unsigned char SHIP_MOVE_SPEED = 2;
 constexpr unsigned char RELOAD_DURATION = 75;
 constexpr unsigned int ALIEN_CHANGE = 200;
-constexpr unsigned int ALIEN_SPEED = 15;
+constexpr float ALIEN_SPEED = 15;
 constexpr unsigned char OFFSET = 20;
 
 #endif //SPACEINVADERS_GLOBAL_H
